Moved files to the home package
Might make reusable components since we need something similar for the characters list :hmm:.
